+++ src/CancellablePromiseInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace React\Promise;
+
+interface CancellablePromiseInterface
+{
+    public function cancel();
+}