+++ tests/PromiseTest/ProgressTestTrait.php
@@ -7,7 +7,7 @@ trait ProgressTestTrait
     /**
      * @return \React\Promise\PromiseAdapter\PromiseAdapterInterface
      */
-    abstract public function getPromiseTestAdapter();
+    abstract public function getPromiseTestAdapter(callable $canceller = null);
 
     /** @test */
     public function progressShouldProgress()