+++ tests/PromiseTest/PromisePendingTestTrait.php
@@ -7,7 +7,7 @@ trait PromisePendingTestTrait
     /**
      * @return \React\Promise\PromiseAdapter\PromiseAdapterInterface
      */
-    abstract public function getPromiseTestAdapter();
+    abstract public function getPromiseTestAdapter(callable $canceller = null);
 
     /** @test */
     public function thenShouldReturnAPromiseForPendingPromise()
@@ -23,5 +23,13 @@ trait PromisePendingTestTrait
         $adapter = $this->getPromiseTestAdapter();
 
         $this->assertInstanceOf('React\\Promise\\PromiseInterface', $adapter->promise()->then(null, null, null));
+    }
+
+    /** @test */
+    public function cancelShouldReturnNullForPendingPromise()
+    {
+        $adapter = $this->getPromiseTestAdapter();
+
+        $this->assertNull($adapter->promise()->cancel());
     }
 }